Add alias resource artifactory_unmanaged_user
Add new resource artifactory_managed_user

Refactor out common code
